requirements.txt
asgiref==3.8.1
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.2
bleach==6.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
crispy-tailwind==1.0.3
decorator==5.2.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.0
django-appconf==1.1.0
django-compressor==4.5.1
django-crispy-forms==2.4
django-formify==0.0.8
django-viewcomponent==1.0.11
docopt==0.6.2
dotenv==0.9.9
executing==2.2.1
fastjsonschema==2.21.2
gunicorn==23.0.0
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pyzmq==27.1.0
rcssmin==1.1.2
referencing==0.37.0
requests==2.32.5
rjsmin==1.2.2
rpds-py==0.29.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
whitenoise==6.11.0
yarg==0.1.9
